Add end point to persist data in GCP

+++ src/main/java/com/datacollector/demo/mapper/ResultsetMapper.java
@@ -24,7 +24,7 @@ public class ResultsetMapper {
         while (resultSet.next()) {
             employees.add(
                     new Employee(
-                            resultSet.getInt("id"),
+                            resultSet.getLong("id"),
                             resultSet.getString("name")
                     ));
         }
